feat(daemon): report promotion_ms so a cold query stops looking instant
A search that spent 21 s paging a cold index back in and 1 ms scanning
it reported `duration_ms: 1`. True, and badly misleading: the client
cannot see that it just paid twenty seconds, so the expensive case is
invisible exactly when someone asks where the time went. Measured in
the field — the daemon booked ~21 s per query while the MCP response
showed 1 ms and 203 ms.

`SearchResponse` gains `promotion_ms`: time inside
`ensure_warm_for_dispatch`, kept separate from `duration_ms` rather
than folded in, because they answer different questions. `duration_ms`
is "how fast is the index"; `promotion_ms` is "what did I pay to have
an index at all". Merging them would make a warm search look slow;
leaving it out hides the slow one.

Timed unconditionally, not only under `--profile` — one `Instant` pair
on a path that takes tens of seconds whenever it does anything, and a
diagnostic you have to know to ask for is one nobody asks for. The
failure paths report it too: a 30 s timeout that spent most of its
budget promoting is a different diagnosis from one that spent it
scanning.

Surfaced where it gets read: `uffs_search` appends "+ 21000ms index
warm-up" to its result header and carries `promotion_ms` in structured
output; `uffs --profile` prints an "Index warm-up" line. Both silent at
zero, so the steady state is unchanged and the line appears only when
it explains something.

Two files crossed the 800-LOC ceiling on the way. Rather than claim
exceptions: the three early-out paths in `search.rs` were building
near-identical empty `SearchResponse` literals — that duplication is
what made adding a field a four-site edit, so it is now one
`empty_response` helper. `main.rs` gives up its `--profile` renderer to
a `client_profile` sibling, which is a better home for a purely
presentational block regardless of line count.
