+++ crates/uffs-client/src/protocol/tests.rs
@@ -289,6 +289,7 @@ fn search_response_inline_rows_round_trip() {
         total_count: 1,
         records_scanned: 1_000_000,
         duration_ms: 8,
+        promotion_ms: None,
         truncated: false,
         profile: None,
         applied_sorts: vec![SearchSortSpec {
@@ -405,6 +406,7 @@ fn search_response_shmem_blob_round_trip() {
         total_count: 168_295,
         records_scanned: 3_571_389,
         duration_ms: 60,
+        promotion_ms: None,
         truncated: false,
         profile: None,
         applied_sorts: vec![],
@@ -465,6 +467,7 @@ fn search_response_inline_blob_round_trip() {
         total_count: 2,
         records_scanned: 250_000,
         duration_ms: 5,
+        promotion_ms: None,
         truncated: false,
         profile: None,
         applied_sorts: vec![],
@@ -537,6 +540,7 @@ fn search_response_shmem_rows_round_trip() {
         total_count: 250_000,
         records_scanned: 5_000_000,
         duration_ms: 42,
+        promotion_ms: None,
         truncated: false,
         profile: None,
         applied_sorts: vec![],
@@ -878,6 +882,7 @@ fn search_response_with_aggregations_round_trip() {
         total_count: 0,
         records_scanned: 500_000,
         duration_ms: 12,
+        promotion_ms: None,
         truncated: false,
         profile: None,
         applied_sorts: vec![],
